Module yaqb::expression

AST types representing various typed SQL expressions. Almost all types implement either Expression or AsExpression.

The most common expression to work with is a Column. There are various methods that you can call on these, found in expression_methods, or operators which you can find in ops.

Any primitive which implements ToSql will also implement AsExpression, allowing it to be used as an argument to any of the methods described here.

Reexports

pub use self::dsl::*;

pub use self::sql_literal::SqlLiteral;

Modules

aliased

array_comparison

bound

count

dsl
Reexports various top level functions and core extensions that are too generic to export by default. This module exists to conveniently glob import in functions where you need them.

expression_methods

extensions

functions

grouped

helper_types

max

ops

ordering

predicates

sql_literal

Traits

AsExpression
Describes how a type can be represented as an expression for a given type. These types couldn't just implement Expression directly, as many things can be used as an expression of multiple types. (String for example, can be used as either VarChar or Text).

BoxableExpression
Helper trait used when boxing expressions. This exists to work around the fact that Rust will not let us use non-core types as bounds on a trait object (you could not return Box<Expression+NonAggregate>)

Expression

NonAggregate
Marker trait to indicate that an expression does not include any aggregate functions. Used to ensure that aggregate expressions aren't mixed with non-aggregate expressions in a select clause, and that they're never included in a where clause.

SelectableExpression
Indicates that an expression can be selected from a source. The second type argument is optional, but is used to indicate that the right side of a left outer join is nullable, even if it wasn't before.