Module yaqb::types::structs

Structs

PgDate

PgInterval
Intervals in Postgres are separated into 3 parts. A 64 bit integer representing time in microseconds, a 32 bit integer representing number of minutes, and a 32 bit integer representing number of months. This struct is a dumb wrapper type, meant only to indicate the meaning of these parts.

PgTime
Time is represented in Postgres as a 64 bit signed integer representing the number of microseconds since midnight. This struct is a dumb wrapper type, meant only to indicate the integer's meaning.

PgTimestamp